requirements.txt
Flask==3.1.0
flask-mail==0.10.0
Flask-Limiter==3.9.0
mysql-connector-python==9.1.0
qrcode[pil]==8.0
opencv-python==4.10.0.84
pyzbar==0.1.9
python-dotenv==1.0.1
fpdf==1.7.2
python-docx==1.1.2
openpyxl==3.1.5
cryptography==44.0.0
email-validator==2.2.0
requests==2.32.3
numpy==2.1.3
pillow==11.0.0
pytest==8.3.3
pytest-flask==1.3.0
pyotp==2.9.0
redis==5.0.1
structlog==24.1.0
python-dateutil==2.8.2
